third_party/blink/web_tests/platform/win11-arm64/http/tests/xmlhttprequest/chromium/access-control-preflight-data-saver-expected.txt

CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/chromium/resources/access-control-preflight-request-invalid-status.php?code=400' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Test that 'Save-Data' header is treated as a simple header and preflight request is not created. Should print PASS.

FAIL
